backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
openai-whisper
pydantic
reportlab
pyyaml
